Validated ID
validated and sanitized all the GETs and POSTs

+++ admin/users/delete_users.php
@@ -7,8 +7,11 @@ include_once('../../includes/connect.php');
 
     if(isset($_SESSION['logged_in'])){
         if(isset($_GET['id'])){
-
-            $id = $_GET['id'];
+            $id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
+            if (!$id || empty($id)) {
+                echo '<script>alert("Incorrect ID passed!!!")</script>';
+                exit;
+            }
 
             $query = $db->prepare("DELETE FROM user WHERE user_id = ?");
             $query->bindValue(1, $id);
